#59 Fixes launchIntercep calculation done with doubles.
Adds CombatCoordinate JUnit test. Style fixes for CombatCoordinate file.

+++ src/main/java/org/openRealmOfStars/player/combat/Combat.java
@@ -240,8 +240,8 @@ public boolean launchIntercept(final int distance,
       double interceptY = shooter.getY();
       double dx, dy;
       if (distance > 0) {
-          dx = (target.getX() - shooter.getX()) / distance;
-          dy = (target.getY() - shooter.getY()) / distance;
+          dx = (double) (target.getX() - shooter.getX()) / distance;
+          dy = (double) (target.getY() - shooter.getY()) / distance;
       } else {
           dx = 0;
           dy = 0;
